requirements.txt
transformers
fastBPE
fairseq
pytorch-crf
vncorenlp